requirements.txt
aiogram>=3.8.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
structlog>=24.4.0
pytest>=8.3.3
pytest-asyncio>=0.24.0
